1.0.0-beta.9: separate entry files for nodejs and browser
